Store archetypes as an array inside filter. Remove extra statics. Add layout info to entity. Improve sparseset enumerator performance. Change license year.

+++ Scellecs.Morpeh/Core/Native/NativeWorld.cs
@@ -6,7 +6,7 @@ namespace Scellecs.Morpeh.Native {
         public int identifier;
         
         [NativeDisableUnsafePtrRestriction]
-        public unsafe int* entitiesGens;
+        public unsafe ushort* entitiesGens;
 
         [NativeDisableUnsafePtrRestriction]
         public unsafe int* entitiesCapacity;
